requirements.txt
python-telegram-bot==22.7
apscheduler==3.11.2
